+++ src/core/annotation.js
@@ -34,7 +34,6 @@ import {
   Util,
   warn,
 } from "../shared/util.js";
-import { Catalog, FileSpec } from "./obj.js";
 import { collectActions, getInheritableProperty } from "./core_utils.js";
 import {
   createDefaultAppearance,
@@ -49,7 +48,9 @@ import {
   Name,
   RefSet,
 } from "./primitives.js";
+import { Catalog } from "./obj.js";
 import { ColorSpace } from "./colorspace.js";
+import { FileSpec } from "./file_spec.js";
 import { ObjectLoader } from "./object_loader.js";
 import { OperatorList } from "./operator_list.js";
 import { StringStream } from "./stream.js";
